+++ samples/Game.Godot.Chat/Server/Hotfix/Chat/ChatService.cs
@@ -17,7 +17,9 @@ namespace Server.Hotfix.Chat
             await call.GameServer.BindCurrentSessionAsync(
                 call.ConnectionId,
                 call.Callback);
-            await call.Actors.AskAsync<ChatRoomActor, bool>(
+            var localActors = call.Actors;
+
+            await localActors.AskAsync<ChatRoomActor, bool>(
                 RoomId,
                 (room, ct) =>
                 {
@@ -28,7 +30,9 @@ namespace Server.Hotfix.Chat
 
         public static async ValueTask SendAsync(HotfixServiceCall<ChatSendRequest, IChatCallback> call)
         {
-            await call.Actors.AskAsync<ChatRoomActor, bool>(
+            var localActors = call.Actors;
+
+            await localActors.AskAsync<ChatRoomActor, bool>(
                 RoomId,
                 (room, ct) =>
                 {
@@ -36,7 +40,7 @@ namespace Server.Hotfix.Chat
                     return new ValueTask<bool>(true);
                 });
             var text = FilterMessage(call.Request.Text ?? "");
-            await call.Actors.AskAsync<ChatRoomActor, bool>(
+            await localActors.AskAsync<ChatRoomActor, bool>(
                 RoomId,
                 async (room, ct) =>
                 {